[Backend] Refactored route to get a list of all officers of a force
