+++ Backend/src/api/officers/index.ts
@@ -2,9 +2,13 @@ import express from 'express';
 import infoRoutes from './info';
 import manageRoutes from './manage';
 import activityRoutes from './activity';
-import {officerExistsMiddle} from "./officer-exists-middle";
+import officerExistsMiddle from "../../middlewares/officer-exists";
+import {getOfficersListController} from "./controllers";
 
 const app = express.Router();
+
+// Route to get a list of all existing officers, following optional filters
+app.get("/", getOfficersListController);
 
 // Load the info routes
 app.use(infoRoutes);
