+++ Backend/src/middlewares/officer-exists.ts
@@ -1,10 +1,10 @@
 import {Request, NextFunction} from "express";
-import {queryDB} from "../../utils/db-connector";
-import {FORCE_HEADER} from "../../utils/constants";
-import {APIResponse} from "../../types/";
+import {queryDB} from "../utils/db-connector";
+import {FORCE_HEADER} from "../utils/constants";
+import {APIResponse} from "../types";
 
 
-export async function officerExistsMiddle(req: Request, res: APIResponse, next: NextFunction) {
+async function officerExistsMiddle(req: Request, res: APIResponse, next: NextFunction) {
     let officerResult = await queryDB(req.header(FORCE_HEADER)!, `SELECT * FROM ${req.query.hasOwnProperty("pretty") ? "officersV" : "officers"} WHERE nif = ?`, req.params.nif);
     if (officerResult.length === 0) {
         res.status(404).json({
@@ -21,3 +21,5 @@ export async function officerExistsMiddle(req: Request, res: APIResponse, next: 
     res.locals.requestedOfficerData.promotion_date = res.locals.requestedOfficerData.promotion_date !== null ? String(new Date(res.locals.requestedOfficerData.promotion_date.getTime() - (res.locals.requestedOfficerData.promotion_date.getTimezoneOffset() * 60000)).toISOString().split("T")[0]): null;
     next();
 }
+
+export default officerExistsMiddle;
